PDP11: Fix TVSEL register.
It's sometimes used as a 16-bit register.

+++ PDP11/pdp11_tv.c
@@ -53,7 +53,7 @@ static uint16 COLORA;
 static uint8  tv_source[256];
 static uint8  tv_display[256];
 static uint16 TVINCR;
-static uint8  TVSEL;     /* Frame buffer select. */
+static uint16 TVSEL;     /* Frame buffer select. */
 static uint16 TVRADR;
 static uint16 tvdata;
 static uint16 TVWC;
